Refactor: deleting all files in any system

+++ src/server.ts
@@ -53,8 +53,6 @@ app.post('/export', (request, response) => {
     fieldDelimiter: ';',
   });
 
-  console.log(path);
-
   csvWriter
     .writeRecords(options)
     .then(() => {
@@ -69,15 +67,16 @@ app.post('/export', (request, response) => {
 });
 
 app.delete('/files', (request, response) => {
-  const { path } = request.body;
+  const path = join(__dirname, '..', '/temp/');
   fs.readdir(path, (err, files) => {
     if (err) throw err;
 
     for (const file of files) {
-      fs.unlinkSync(join(path, file));
+      if (file.includes('.csv')) {
+        fs.unlinkSync(join(path, file));
+      }
     }
   });
-  // fs.unlinkSync(`${path}/Teste_enviar_envs.csv`);
 
   return response.send();
 });
